feat: create show date picker
- Create a date picker for the next seven days. Note that the schedule
for next Fri to Thu is released _around_ 15.00 on Tuesday. As a
result, the component sometimes shows dates for which the schedule has
not yet been published. This should be addressed by showing a
descriptive message if no shows are found. I chose to always show the
next seven days instead of creating some heuristic because the schedule
from Finnkino is not exact. Also, this was easier:)
- Use focus-visible polyfill to only show focus styles when needed for
a11y, e.g. not on mouse navigation.